Fix optimal dictionary compression
